# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register new user",
                "description": "User registration request",
                "operationId": "registerUser",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Token for registration",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJpd...aWLuQ3lVXs1A"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "position_id",
                                    "photo"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Username should contain 2-60 characters.",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email, must be a valid email according to RFC2822.",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "User phone number. Number should start with code of Ukraine +380.",
                                        "type": "string"
                                    },
                                    "position_id": {
                                        "description": "User's position ID.",
                                        "type": "integer"
                                    },
                                    "photo": {
                                        "description": "Minimum size of photo 70x70px. The photo format must be jpeg/jpg type. The photo size must not be greater than 5 Mb.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "New user successfully registered"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Expired token response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The token expired."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}